Added Foldouts to remove visual clutter
ProjectSection Foldout is enabled by default.

Fixed Argument Flag issue where you couldn't set the button correctly.

A "TSKDoNotDelete" text file is generated when the package is first added, this was done to counter the constant auto-generation that ocurred when you would delete the config files.

Updated .gitignore to not commit TSKDoNotDelete.txt